requirements.txt
python-dotenv==1.0.0
requests==2.31.0
solana==0.30.2
solders==0.18.0
colorlog==6.7.0
